move class to own file
